+++ l10n/fr.js
@@ -116,7 +116,6 @@ OC.L10N.register(
     "Create new Form template": "Créer un nouveau modèle de formulaire",
     "Create form": "Créer un formulaire",
     "Fill in form in ONLYOFFICE": "Remplir le formulaire dans ONLYOFFICE",
-    "Please update ONLYOFFICE Docs to version 7.0 to work on fillable forms online": "Veuillez mettre à jour ONLYOFFICE Docs vers la version 7.0 pour travailler sur les formulaires à remplir en ligne",
     "Security": "Sécurité",
     "Run document macros": "Exécuter des macros de documents",
     "Anonymous": "Anonyme",