+++ lib/Controller/SettingsController.php
@@ -1,7 +1,7 @@
 <?php
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * This program is a free software product.
  * You can redistribute it and/or modify it under the terms of the GNU Affero General Public License
@@ -127,6 +127,7 @@ class SettingsController extends Controller {
             "sameTab" => $this->config->getSameTab(),
             "preview" => $this->config->getPreview(),
             "advanced" => $this->config->getAdvanced(),
+            "cronChecker" => $this->config->getCronChecker(),
             "versionHistory" => $this->config->getVersionHistory(),
             "protection" => $this->config->getProtection(),
             "limitGroups" => $this->config->getLimitGroups(),
@@ -213,6 +214,7 @@ class SettingsController extends Controller {
      * @param bool $sameTab - open in the same tab
      * @param bool $preview - generate preview files
      * @param bool $advanced - use advanced tab
+     * @param bool $cronChecker - disable cron checker
      * @param bool $versionHistory - keep version history
      * @param array $limitGroups - list of groups
      * @param bool $chat - display chat
@@ -231,6 +233,7 @@ class SettingsController extends Controller {
         $sameTab,
         $preview,
         $advanced,
+        $cronChecker,
         $versionHistory,
         $limitGroups,
         $chat,
@@ -248,6 +251,7 @@ class SettingsController extends Controller {
         $this->config->setSameTab($sameTab);
         $this->config->setPreview($preview);
         $this->config->setAdvanced($advanced);
+        $this->config->setCronChecker($cronChecker);
         $this->config->setVersionHistory($versionHistory);
         $this->config->setLimitGroups($limitGroups);
         $this->config->setCustomizationChat($chat);