+++ src/viewer.js
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * This program is a free software product.
  * You can redistribute it and/or modify it under the terms of the GNU Affero General Public License
@@ -86,7 +86,7 @@
                 return format.mime;
             }
         });
-
+        mimes.flat();
         OCA.Viewer.registerHandler({
             id: OCA.Onlyoffice.AppName,
             group: null,