+++ templates/settings.php
@@ -1,7 +1,7 @@
 <?php
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * This program is a free software product.
  * You can redistribute it and/or modify it under the terms of the GNU Affero General Public License
@@ -139,6 +139,12 @@ if ($_["tagsEnabled"]) {
         <button id="onlyofficeClearVersionHistory" class="button"><?php p($l->t("Clear")) ?></button>
     </p>
 
+    <p>
+        <input type="checkbox" class="checkbox" id="onlyofficeCronChecker"
+            <?php if ($_["cronChecker"]) { ?>checked="checked"<?php } ?> />
+        <label for="onlyofficeCronChecker"><?php p($l->t("Enable background connection check to the editors")) ?></label>
+    </p>
+
     <p class="onlyoffice-header"><?php p($l->t("The default application for opening the format")) ?></p>
     <div class="onlyoffice-exts">
         <?php foreach ($_["formats"] as $format => $setting) { ?>
@@ -160,7 +166,7 @@ if ($_["tagsEnabled"]) {
     </p>
     <div class="onlyoffice-exts">
         <?php foreach ($_["formats"] as $format => $setting) { ?>
-            <?php if (array_key_exists("editable", $setting)) { ?>
+            <?php if (array_key_exists("editable", $setting) && $setting["editable"]) { ?>
             <div>
                 <input type="checkbox" class="checkbox"
                     id="onlyofficeEditFormat<?php p($format) ?>"